% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacer.R
\name{adjacer}
\alias{adjacer}
\title{Adjacer}
\usage{
adjacer(my, mx, r, prob, desnare.reps)
}
\arguments{
\item{my}{slice}

\item{mx}{all raster cells (with predicted snare occurrence and not entirely open areas)}

\item{r}{raster from which mx object was obtained}

\item{prob}{probability of detecting a snare}

\item{desnare.reps}{the number of repeats in surrounding snare cells}
}
\value{
slice with adjacent cells added, including whether snares were found (snares.repeat) and updated counter (count.repeat)
}
\description{
Takes a slice of cells and identifies adjacent cells.
}
