% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{: m}
\alias{: m}
\alias{m}
\title{Point pattern points of snares}
\format{data.table}
\source{
master.R
}
\usage{
m
}
\description{
The package "spatstat" requires its own format ("ppp").
The number of snares was added as a mark.
See all_snares for a ppp with unique snare locations
}
\keyword{raster}
\keyword{snares,}
