% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desnare_adjacent.R
\name{desnare_adjacent}
\alias{desnare_adjacent}
\title{desnare cells adjacent to snare cells}
\usage{
desnare_adjacent(mx, include.bush = F, desnare.reps = 2, prob = 0.4)
}
\arguments{
\item{mx}{output of function `mx`}

\item{desnare.reps}{how many times desnaring must be repeated for adjacent cells (a repeat is a proxy for spending more time in the cell for desnaring).}

\item{prob}{probability of finding a snare}

\item{bush}{whether to include bush into the adjacent cells}
}
\description{
using the output of function `sequential`: identify snare cells (where you found cells during the sequential desnaring run);
for each of these cells, find the adjacent cells (Moore neighbourhood).
use only cells that you did not previously desnare, where there is bush and where there was predicted occurrence of snares.
repeat desnaring n=desnare.reps times, and update counts, snares found.
}
\examples{
mx=sequential(m); l=desnare_adjacent(mx)
}
