% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_slicer.R
\name{hotspot_slicer}
\alias{hotspot_slicer}
\title{hotspot search}
\usage{
hotspot_slicer(mx, r, i)
}
\arguments{
\item{mx}{cell overview with bush>0 and predicted snares prob > 0 (= mx object)}

\item{r}{raster with snares or predictions}

\item{i}{index number of the pass in the slicer function, of which this function is part}
}
\value{
my slice with label 'hotspot', based on sample of snare memory and adjacent cells
}
\description{
takes a sample of adjacent cells
}
