% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sequential_adjacent.R
\name{merge_sequential_adjacent}
\alias{merge_sequential_adjacent}
\title{in object z, .id comes from the repeats. so, in z[found>0,] non-NA values for .id means that the snares were found during repeat desnaring,
if they are NA, then they were found during the first round of desnaring.
id comes from the snare cell summary data.table sdf. here, the id stands for the snare cell, in order of identification.
here, id indicates that the snare was found during the first sequential desnaring, and NA that they were found during the repeats.}
\usage{
merge_sequential_adjacent(mx, l)
}
\description{
in object z, .id comes from the repeats. so, in z[found>0,] non-NA values for .id means that the snares were found during repeat desnaring,
if they are NA, then they were found during the first round of desnaring.
id comes from the snare cell summary data.table sdf. here, the id stands for the snare cell, in order of identification.
here, id indicates that the snare was found during the first sequential desnaring, and NA that they were found during the repeats.
}
