% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{mz}
\alias{mz}
\title{Sorted data.table of raster values}
\format{data.table}
\source{
prepare_data.R
}
\usage{
mz
}
\description{
Table with predictions per cell, covariates and cell positions
}
\keyword{m}
