% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{plot results}
\usage{
plot_results(
  m,
  prob,
  sims,
  r,
  perc = T,
  onepage = T,
  printpdf = F,
  reps,
  magn = 2
)
}
\arguments{
\item{m}{cell overview}

\item{prob}{Probability of detection a snare}

\item{sims}{data.table with simulation result}

\item{r}{raster with snares, or any other raster that has same extent and resolution}

\item{perc}{Whether you plot raw snare counts or percentage of discoverable snares}

\item{onepage}{whether or not to plot summary plots on one page}

\item{printpdf}{whether or not to print a pdf of the performance plot}

\item{reps}{the number of repetitions in the simulation}

\item{magn}{the amount of magnification (cex) for plot text, symbols and axes}
}
\value{
plot of stabilization of results, plot of cumulative result over number of cells visited,
discoverable, found and non-discoverable snares, average visits per cell
}
\description{
Plot the results of a simulation.
}
\examples{
sim=foreach(i=1:reps) \%dopar\% {
sequential_desnare(m)
}
sims=rbindlist(sim,idcol=T)
plot_results(snares.pred,0.4,sims,perc=F)
}
