% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotclusters.r
\name{plotclusters}
\alias{plotclusters}
\title{Plot clusters}
\usage{
plotclusters(p, distance, add = F, clusterpoints = 5, main = NULL, plot = T)
}
\arguments{
\item{p}{input spatstat ppp object}

\item{distance}{the distance that you want to apply to the dilation (Steiner set), see ?connected.ppp}

\item{add}{Whether you want to add the pattern to an existing map or not. Default is FALSE.}

\item{clusterpoints}{the number of objects (e.g. snares) that you want to have in one cluster. Default is 5.}
}
\value{
A list of hotspot characteristics and a hotspot plot. hotspots A list of individual hotspots within the overall window. hotspot_centroids The centroids of the hotspots that you found as SpatialPoints class object. raster_hotspots The hotspots as a raster object
}
\description{
Find clusters of snares based on connectivity (dilation; Steiner set) and the minimum number of snares that you want to be in that hotspot.
The default dilation distance is based on the 90% quantile of the nearest distance of snares for k=2 (the second nearest snare).
}
\details{
This method was suggested by one of the authors of the spatstat package, Ege Rubak on my stackoverflow question:
https://stackoverflow.com/questions/55656101/isolate-singletons-in-groups-using-stienen-sets-and-dilation/55680507#55680507
}
\examples{
plotstuff(add=F, legend=F, cex=0.8, lwd=2)
plotclusters(snares, clusterpoints=5)
}
