% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnare.R
\name{resnare}
\alias{resnare}
\title{resnare}
\usage{
resnare(sim, m, r, n.clumps, resnare.frac, cells.only = T)
}
\arguments{
\item{sim}{data.table of desnaring simulation of a single run}

\item{m}{overview of raster cells with snares}

\item{r}{raster overview of snares (or any raster with same extent and resolution)}

\item{n.clumps}{number of snare clumps to resnare}

\item{resnare.frac}{fraction of snare cells that will be replaced (if they were removed during desnaring)}

\item{cells.only}{whether to return contagion cells (default=TRUE) or whether the whole sim object is returned.}
}
\value{
simulation object with snares replaced for selected cells. `recent` field will be set to 1
}
\description{
Takes a simulation run of desnaring and re-snares.
Take cells with snares in them and surrounding cells that have some bush in them.
Find clumps and select some of them through taking a sample.
Within the selected clumps, re-snare a fraction of the cells.
Both clump and cell selection have seeds set, so the selected cells are always the same in view of reproducibility.
The default is to return only cells, since this is much less computationally intensive.
}
\examples{
resnare(sim,m,r,n.clumps=3,resnare.frac=1)
}
