% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resnare_update.R
\name{resnare_update}
\alias{resnare_update}
\title{Update run with resnares}
\usage{
resnare_update(x, prob)
}
\arguments{
\item{x}{the result of running the function slicer with sequential.only=F}

\item{prob}{the probability of detecting a snare.}
}
\value{
x the updated result of the slicer function output. Where resnared cells were discovered, the found.rep field was increased with the recent field.
}
\description{
You have run a desnaring exercise with search of adjacent cells if you found a snare (slicer with sequential.only=F).
Some cells were resnared (recent=1). These cells may have been visited later on, if repeats > 0.
}
\details{
Here, you will filter out these cells and do a repeat desnaring.
(The function is not recursive: in reality, if you would find a snare, you would search the adjacent cells, which could lead
to finding new snares, which could lead to search of adjacent cells etc.)

If you find a snare, update the found.rep field and the found field.
}
\examples{
x=slicer(m,r,prob=0.4,n.sample=20,desnare.reps=2,sequential.only=F,cutoff=T)
y=resnare_update(x,prob=0.4)
}
