% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_desnare.R
\name{sequential_desnare}
\alias{sequential_desnare}
\title{Perform a one-pass sequential desnaring operation over a data.table}
\usage{
sequential_desnare(m, prob = 0.4)
}
\arguments{
\item{m}{data.table with raster cells, coordinates and snares (data(m))}

\item{prob}{probability of detecting snares, default=0.4}
}
\value{
mx reduced object m (no bush, no cells for which no snare occurrence was predicted). Fields added: counter, found.
}
\description{
A raster with snares found during testing a snare occurrence model was made and processed, and available as object m.
Here, we remove raster cells for which there is no prediction of snares, and also no bush (nothing to attach the snares to).
A dice is thrown for each row (cell) of the object; if it is smaller than the probability of finding the snare `prob`, then the snare is found.
In that case, the `snare` field is reset to zero, and `found` to 1.
}
\examples{
sequential(m,prob=0.2)
}
