% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_cutoff.R
\name{sort_cutoff}
\alias{sort_cutoff}
\title{Sort and cutoff}
\usage{
sort_cutoff(z)
}
\arguments{
\item{z}{data.table after doing sequential and adjacent desnare and merge of the results}
}
\value{
data.table cutoff at an effort equal to the number of cells visited during sequential desnare
}
\description{
Sort and cutoff
}
